+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `goesaodc_createScanPoints()` function now returns a list of 
 `SpatialPointsDataFrames` when multiple scans are requested.
+* `goesaodc_createScanRaster()` function now returns a list of 
+`RasterBricks` when multiple scans are requested.
 * Removed 'Kincade Fire Example' article.
 * Removed `createSpatialPointsVideo_exec` script.
 * Removed `goesaodc_areaPlot()` function.